JSFile : Fix crash in JS_ReportError caused by unknown format character

+++ src/Binding/JSFile.cpp
@@ -873,12 +873,12 @@ static bool nidium_file_seekSync(JSContext *cx, unsigned argc, JS::Value *vp)
 
     if (ret != 0) {
         if (err == 0) {
-            JS_ReportError(cx, "Unable to seek to %zu : %s", 
-                    (size_t)seekPos, !file->isOpen() ? "not opened" : "is it a directory?");
+            JS_ReportError(cx, "Unable to seek to %f : %s",
+                    seekPos, (!file->isOpen() ? "not opened" : "is it a directory?"));
             return false;
         } else {
-            JS_ReportError(cx, "Failed to seek to %zu : %s (errno %d)", 
-                    (size_t)seekPos, strerror(err), err);
+            JS_ReportError(cx, "Failed to seek to %f : %s (errno %d)",
+                    seekPos, strerror(err), err);
             return false;
         }
     }
